docker/requirements.txt
biopython==1.72
DendroPy==4.4.0
numpy==1.14.5
pandas==0.23.2
scipy==1.1.0
statsmodels==0.9.0
